backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
selenium==4.15.2
pytest==7.4.3
pytest-html==4.1.1
google-generativeai==0.3.1
python-dotenv==1.0.0

